+++ OurUmbraco/OurUmbraco.csproj
@@ -420,6 +420,7 @@
     <Compile Include="CamelCaseFormatter.cs" />
     <Compile Include="Community\BlogPosts\BlogCachedRssItem.cs" />
     <Compile Include="Community\BlogPosts\BlogInfo.cs" />
+    <Compile Include="Community\BlogPosts\BlogPostsCache.cs" />
     <Compile Include="Community\BlogPosts\BlogPostsService.cs" />
     <Compile Include="Community\BlogPosts\BlogRssChannel.cs" />
     <Compile Include="Community\BlogPosts\BlogRssItem.cs" />
